+++ apps/vite-react-project/.eslintrc.cjs
@@ -1,6 +1,6 @@
 module.exports = {
   root: true,
-  extends: ["@mocha-design/eslint-config"],
+  extends: ["@coconut-killer/eslint-config"],
   // plugins: ["react-refresh"],
   // rules: {
   //   "react-refresh/only-export-components": [