+++ apps/vite-react-project/src/App.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import './App.css'
-import { Button } from "@mocha-design/ui/react";
+import { Button } from "@coconut-killer/ui/react";
 
 function App() {
   const [count, setCount] = useState(0)